build(py): Restore Python-specific template bits
This reverts commit ea0ca8b3b9ca0a1469a0f7da4742f631016a2893.

+++ Makefile
@@ -8,6 +8,8 @@
 
 # Variables used as options to control behavior:
 export TEMPLATE_IGNORE_EXISTING=false
+# https://devguide.python.org/versions/#supported-versions
+PYTHON_SUPPORTED_MINORS=3.11 3.10 3.9 3.8 3.7
 
 
 ## "Private" Variables:
@@ -41,6 +43,36 @@ ifeq ($(USER_FULL_NAME),)
 USER_FULL_NAME=$(USER_NAME)
 endif
 USER_EMAIL:=$(USER_NAME)@$(shell hostname -f)
+
+# Values concerning supported Python versions:
+# Use the same Python version tox would as a default.
+# https://tox.wiki/en/latest/config.html#base_python
+PYTHON_HOST_MINOR:=$(shell \
+    pip --version | sed -nE 's|.* \(python ([0-9]+.[0-9]+)\)$$|\1|p;q')
+export PYTHON_HOST_ENV=py$(subst .,,$(PYTHON_HOST_MINOR))
+# Determine the latest installed Python version of the supported versions
+PYTHON_BASENAMES=$(PYTHON_SUPPORTED_MINORS:%=python%)
+PYTHON_AVAIL_EXECS:=$(foreach \
+    PYTHON_BASENAME,$(PYTHON_BASENAMES),$(shell which $(PYTHON_BASENAME)))
+PYTHON_LATEST_EXEC=$(firstword $(PYTHON_AVAIL_EXECS))
+PYTHON_LATEST_BASENAME=$(notdir $(PYTHON_LATEST_EXEC))
+PYTHON_MINOR=$(PYTHON_HOST_MINOR)
+ifeq ($(PYTHON_MINOR),)
+# Fallback to the latest installed supported Python version
+PYTHON_MINOR=$(PYTHON_LATEST_BASENAME:python%=%)
+endif
+PYTHON_LATEST_MINOR=$(firstword $(PYTHON_SUPPORTED_MINORS))
+PYTHON_LATEST_ENV=py$(subst .,,$(PYTHON_LATEST_MINOR))
+PYTHON_MINORS=$(PYTHON_SUPPORTED_MINORS)
+ifeq ($(PYTHON_MINOR),)
+PYTHON_MINOR=$(firstword $(PYTHON_MINORS))
+else ifeq ($(findstring $(PYTHON_MINOR),$(PYTHON_MINORS)),)
+PYTHON_MINOR=$(firstword $(PYTHON_MINORS))
+endif
+export PYTHON_ENV=py$(subst .,,$(PYTHON_MINOR))
+PYTHON_SHORT_MINORS=$(subst .,,$(PYTHON_MINORS))
+PYTHON_ENVS=$(PYTHON_SHORT_MINORS:%=py%)
+PYTHON_ALL_ENVS=$(PYTHON_ENVS) build
 
 # Values derived from VCS/git:
 VCS_LOCAL_BRANCH:=$(shell git branch --show-current)
@@ -111,14 +143,23 @@ ifneq ($(VCS_MERGE_BRANCH),$(VCS_BRANCH))
 VCS_FETCH_TARGETS+=./var/git/refs/remotes/$(VCS_REMOTE)/$(VCS_MERGE_BRANCH)
 endif
 
-# Run Python tools in isolated environments managed by Tox:
+# Values used to run Tox:
+TOX_ENV_LIST=$(subst $(EMPTY) ,$(COMMA),$(PYTHON_ENVS))
+TOX_RUN_ARGS=run-parallel --parallel auto --parallel-live
+ifeq ($(words $(PYTHON_MINORS)),1)
+TOX_RUN_ARGS=run
+endif
+# The options that allow for rapid execution of arbitrary commands in the venvs managed
+# by tox
 TOX_EXEC_OPTS=--no-recreate-pkg --skip-pkg-install
+TOX_EXEC_ARGS=tox exec $(TOX_EXEC_OPTS) -e "$(PYTHON_ENV)"
 TOX_EXEC_BUILD_ARGS=tox exec $(TOX_EXEC_OPTS) -e "build"
 
 # Values used for publishing releases:
 # Safe defaults for testing the release process without publishing to the final/official
 # hosts/indexes/registries:
 RELEASE_PUBLISH=false
+PYPI_REPO=testpypi
 # Only publish releases from the `main` or `develop` branches:
 ifeq ($(VCS_BRANCH),main)
 RELEASE_PUBLISH=true
@@ -126,6 +167,14 @@ else ifeq ($(VCS_BRANCH),develop)
 # Publish pre-releases from the `develop` branch:
 RELEASE_PUBLISH=true
 endif
+ifeq ($(RELEASE_PUBLISH),true)
+PYPI_REPO=pypi
+endif
+# Address undefined variables warnings when running under local development
+PYPI_PASSWORD=
+export PYPI_PASSWORD
+TEST_PYPI_PASSWORD=
+export TEST_PYPI_PASSWORD
 
 # Done with `$(shell ...)`, echo recipe commands going forward
 .SHELLFLAGS+= -x
@@ -156,12 +205,36 @@ all: build
 ### Perform any currently necessary local set-up common to most operations.
 build: \
 	./.git/hooks/pre-commit \
-	$(HOME)/.local/var/log/project-structure-host-install.log
+	$(HOME)/.local/var/log/python-project-structure-host-install.log
+	$(MAKE) -e -j $(PYTHON_ENVS:%=build-requirements-%)
+
+.PHONY: $(PYTHON_ENVS:%=build-requirements-%)
+### Compile fixed/pinned dependency versions if necessary.
+$(PYTHON_ENVS:%=build-requirements-%):
+# Avoid parallel tox recreations stomping on each other
+	$(MAKE) -e "$(@:build-requirements-%=./var/log/tox/%/build.log)"
+	targets="./requirements/$(@:build-requirements-%=%)/user.txt \
+	    ./requirements/$(@:build-requirements-%=%)/devel.txt \
+	    ./requirements/$(@:build-requirements-%=%)/build.txt \
+	    ./build-host/requirements-$(@:build-requirements-%=%).txt"
+# Workaround race conditions in pip's HTTP file cache:
+# https://github.com/pypa/pip/issues/6970#issuecomment-527678672
+	$(MAKE) -e -j $${targets} ||
+	    $(MAKE) -e -j $${targets} ||
+	    $(MAKE) -e -j $${targets}
 
 .PHONY: build-pkgs
 ### Ensure the built package is current when used outside of tox.
 build-pkgs: ./var/git/refs/remotes/$(VCS_REMOTE)/$(VCS_BRANCH)
-	true "TEMPLATE: Always specific to the type of project"
+# Defined as a .PHONY recipe so that multiple targets can depend on this as a
+# pre-requisite and it will only be run once per invocation.
+	rm -vf ./dist/*
+	tox run -e "$(PYTHON_ENV)" --pkg-only
+# Copy the wheel to a location not managed by tox:
+	cp -lfv "$$(ls -t ./.tox/.pkg/dist/*.whl | head -n 1)" "./dist/"
+# Also build the source distribution:
+	tox run -e "$(PYTHON_ENV)" --override "testenv.package=sdist" --pkg-only
+	cp -lfv "$$(ls -t ./.tox/.pkg/dist/*.tar.gz | head -n 1)" "./dist/"
 
 
 ## Test Targets:
@@ -171,23 +244,17 @@ build-pkgs: ./var/git/refs/remotes/$(VCS_REMOTE)/$(VCS_BRANCH)
 .PHONY: test
 ### Run the full suite of tests, coverage checks, and linters.
 test: build
-	true "TEMPLATE: Always specific to the type of project"
-
-.PHONY: test-lint
-### Perform any linter or style checks, including non-code checks.
-test-lint: $(HOME)/.local/var/log/project-structure-host-install.log
-# Run non-code checks, e.g. documentation:
-	tox run -e "build"
+	tox $(TOX_RUN_ARGS) -e "$(TOX_ENV_LIST)"
 
 .PHONY: test-debug
 ### Run tests directly on the host and invoke the debugger on errors/failures.
-test-debug:
-	true "TEMPLATE: Always specific to the type of project"
+test-debug: ./var/log/tox/$(PYTHON_ENV)/editable.log
+	$(TOX_EXEC_ARGS) -- pytest --pdb
 
 .PHONY: test-push
 ### Perform any checks that should only be run before pushing.
 test-push: $(VCS_FETCH_TARGETS) \
-		$(HOME)/.local/var/log/project-structure-host-install.log
+		$(HOME)/.local/var/log/python-project-structure-host-install.log
 	vcs_compare_rev="$(VCS_COMPARE_REMOTE)/$(VCS_COMPARE_BRANCH)"
 	if ! git fetch "$(VCS_COMPARE_REMOTE)" "$(VCS_COMPARE_BRANCH)"
 	then
@@ -228,19 +295,24 @@ test-clean:
 # end-users.
 
 .PHONY: release
-### Publish installable packages if conventional commits require a release.
-release: $(HOME)/.local/var/log/project-structure-host-install.log
+### Publish installable Python packages if conventional commits require a release.
+release: $(HOME)/.local/var/log/python-project-structure-host-install.log ~/.pypirc
 # Only release from the `main` or `develop` branches:
 ifeq ($(RELEASE_PUBLISH),true)
 	$(MAKE) -e build-pkgs
-	true "TEMPLATE: Always specific to the type of project"
+# https://twine.readthedocs.io/en/latest/#using-twine
+	$(TOX_EXEC_BUILD_ARGS) -- twine check ./dist/python?project?structure-*
+# The VCS remote should reflect the release before the release is published to ensure
+# that a published release is never *not* reflected in VCS.
 	$(MAKE) -e test-clean
+	$(TOX_EXEC_BUILD_ARGS) -- twine upload -s -r "$(PYPI_REPO)" \
+	    ./dist/python?project?structure-*
 endif
 
 .PHONY: release-bump
 ### Bump the package version if on a branch that should trigger a release.
 release-bump: ~/.gitconfig $(VCS_RELEASE_FETCH_TARGETS) \
-		$(HOME)/.local/var/log/project-structure-host-install.log
+		$(HOME)/.local/var/log/python-project-structure-host-install.log
 	if ! git diff --cached --exit-code
 	then
 	    set +x
@@ -304,16 +376,19 @@ endif
 
 .PHONY: devel-format
 ### Automatically correct code in this checkout according to linters and style checkers.
-devel-format: $(HOME)/.local/var/log/project-structure-host-install.log
+devel-format: $(HOME)/.local/var/log/python-project-structure-host-install.log
 	$(TOX_EXEC_ARGS) -- autoflake -r -i --remove-all-unused-imports \
 		--remove-duplicate-keys --remove-unused-variables \
-		--remove-unused-variables "./src/project-structure/"
-	$(TOX_EXEC_ARGS) -- autopep8 -v -i -r "./src/project-structure/"
-	$(TOX_EXEC_ARGS) -- black "./src/project-structure/"
+		--remove-unused-variables "./src/pythonprojectstructure/"
+	$(TOX_EXEC_ARGS) -- autopep8 -v -i -r "./src/pythonprojectstructure/"
+	$(TOX_EXEC_ARGS) -- black "./src/pythonprojectstructure/"
 
 .PHONY: devel-upgrade
 ### Update all fixed/pinned dependencies to their latest available versions.
 devel-upgrade:
+	touch "./setup.cfg" "./requirements/build.txt.in" \
+	    "./build-host/requirements.txt.in"
+	$(MAKE) -e -j $(PYTHON_ENVS:%=build-requirements-%)
 # Update VCS hooks from remotes to the latest tag.
 	$(TOX_EXEC_BUILD_ARGS) -- pre-commit autoupdate
 
@@ -330,12 +405,12 @@ devel-upgrade-branch: ~/.gitconfig ./var/git/refs/remotes/$(VCS_REMOTE)/$(VCS_BR
 	fi
 # Commit the upgrade changes
 	echo "Upgrade all requirements to the latest versions as of $${now}." \
-	    >"./src/projectstructure/newsfragments/\
+	    >"./src/pythonprojectstructure/newsfragments/\
 	+upgrade-requirements.bugfix.rst"
 	git add --update './build-host/requirements-*.txt' './requirements/*/*.txt' \
 	    "./.pre-commit-config.yaml"
 	git add \
-	    "./src/projectstructure/newsfragments/\
+	    "./src/pythonprojectstructure/newsfragments/\
 	+upgrade-requirements.bugfix.rst"
 	git commit --all --gpg-sign -m \
 	    "fix(deps): Upgrade requirements latest versions"
@@ -371,12 +446,67 @@ clean:
 #
 # Recipes that make actual changes and create and update files for the target.
 
+# Manage fixed/pinned versions in `./requirements/**.txt` files.  Has to be run for each
+# python version in the virtual environment for that Python version:
+# https://github.com/jazzband/pip-tools#cross-environment-usage-of-requirementsinrequirementstxt-and-pip-compile
+$(PYTHON_ENVS:%=./requirements/%/devel.txt): ./pyproject.toml ./setup.cfg ./tox.ini
+	true DEBUG Updated prereqs: $(?)
+	$(MAKE) -e "$(@:requirements/%/devel.txt=./var/log/tox/%/build.log)"
+	./.tox/$(@:requirements/%/devel.txt=%)/bin/pip-compile \
+	    --resolver "backtracking" --upgrade --extra "devel" \
+	    --output-file "$(@)" "$(<)"
+$(PYTHON_ENVS:%=./requirements/%/user.txt): ./pyproject.toml ./setup.cfg ./tox.ini
+	true DEBUG Updated prereqs: $(?)
+	$(MAKE) -e "$(@:requirements/%/user.txt=./var/log/tox/%/build.log)"
+	./.tox/$(@:requirements/%/user.txt=%)/bin/pip-compile \
+	    --resolver "backtracking" --upgrade --output-file "$(@)" "$(<)"
+$(PYTHON_ENVS:%=./build-host/requirements-%.txt): ./build-host/requirements.txt.in
+	true DEBUG Updated prereqs: $(?)
+	$(MAKE) -e "$(@:build-host/requirements-%.txt=./var/log/tox/%/build.log)"
+	./.tox/$(@:build-host/requirements-%.txt=%)/bin/pip-compile \
+	    --resolver "backtracking" --upgrade --output-file "$(@)" "$(<)"
+# Only update the installed tox version for the latest/host/main/default Python version
+	if [ "$(@:build-host/requirements-%.txt=%)" = "$(PYTHON_ENV)" ]
+	then
+# Don't install tox into one of it's own virtual environments
+	    if [ -n "$${VIRTUAL_ENV:-}" ]
+	    then
+	        pip_bin="$$(which -a pip | grep -v "^$${VIRTUAL_ENV}/bin/" | head -n 1)"
+	    else
+	        pip_bin="pip"
+	    fi
+	    "$${pip_bin}" install -r "$(@)"
+	fi
+$(PYTHON_ENVS:%=./requirements/%/build.txt): ./requirements/build.txt.in
+	true DEBUG Updated prereqs: $(?)
+	$(MAKE) -e "$(@:requirements/%/build.txt=./var/log/tox/%/build.log)"
+	./.tox/$(@:requirements/%/build.txt=%)/bin/pip-compile \
+	    --resolver "backtracking" --upgrade --output-file "$(@)" "$(<)"
+
+# Targets used as pre-requisites to ensure virtual environments managed by tox have been
+# created and can be used directly to save time on Tox's overhead when we don't need
+# Tox's logic about when to update/recreate them, e.g.:
+#     $ ./.tox/build/bin/cz --help
+# Mostly useful for build/release tools.
+$(PYTHON_ALL_ENVS:%=./var/log/tox/%/build.log):
+	$(MAKE) -e "$(HOME)/.local/var/log/python-project-structure-host-install.log"
+	mkdir -pv "$(dir $(@))"
+	tox run $(TOX_EXEC_OPTS) -e "$(@:var/log/tox/%/build.log=%)" --notest |&
+	    tee -a "$(@)"
+# Workaround tox's `usedevelop = true` not working with `./pyproject.toml`.  Use as a
+# prerequisite when using Tox-managed virtual environments directly and changes to code
+# need to take effect immediately.
+$(PYTHON_ENVS:%=./var/log/tox/%/editable.log):
+	$(MAKE) -e "$(HOME)/.local/var/log/python-project-structure-host-install.log"
+	mkdir -pv "$(dir $(@))"
+	tox exec $(TOX_EXEC_OPTS) -e "$(@:var/log/tox/%/editable.log=%)" -- \
+	    pip install -e "./" |& tee -a "$(@)"
 
 # Install all tools required by recipes that have to be installed externally on the
 # host.  Use a target file outside this checkout to support multiple checkouts.  Use a
 # target specific to this project so that other projects can use the same approach but
 # with different requirements.
-$(HOME)/.local/var/log/project-structure-host-install.log:
+$(HOME)/.local/var/log/python-project-structure-host-install.log:
 	mkdir -pv "$(dir $(@))"
 	(
 	    if ! which pip
@@ -394,7 +524,12 @@ $(HOME)/.local/var/log/project-structure-host-install.log:
 	            sudo apt-get install -y "gettext-base" "python3-pip"
 	        fi
 	    fi
-	    pip install -r "./build-host/requirements.txt.in"
+	    if [ -e ./build-host/requirements-$(PYTHON_HOST_ENV).txt ]
+	    then
+	        pip install -r "./build-host/requirements-$(PYTHON_HOST_ENV).txt"
+	    else
+	        pip install -r "./build-host/requirements.txt.in"
+	    fi
 	) |& tee -a "$(@)"
 
 # Retrieve VCS data needed for versioning (tags) and release (release notes).
@@ -415,9 +550,14 @@ $(VCS_FETCH_TARGETS): ./.git/logs/HEAD
 	fi
 
 ./.git/hooks/pre-commit:
-	$(MAKE) -e "$(HOME)/.local/var/log/project-structure-host-install.log"
+	$(MAKE) -e "$(HOME)/.local/var/log/python-project-structure-host-install.log"
 	$(TOX_EXEC_BUILD_ARGS) -- pre-commit install \
 	    --hook-type "pre-commit" --hook-type "commit-msg" --hook-type "pre-push"
+
+# Capture any project initialization tasks for reference.  Not actually usable.
+./pyproject.toml:
+	$(MAKE) -e "$(HOME)/.local/var/log/python-project-structure-host-install.log"
+	$(TOX_EXEC_BUILD_ARGS) -- cz init
 
 # Tell Emacs where to find checkout-local tools needed to check the code.
 ./.dir-locals.el: ./.dir-locals.el.in
@@ -428,6 +568,10 @@ $(VCS_FETCH_TARGETS): ./.git/logs/HEAD
 	git config --global user.name "$(USER_FULL_NAME)"
 	git config --global user.email "$(USER_EMAIL)"
 
+# Ensure release publishing authentication, mostly useful in automation such as CI.
+~/.pypirc: ./home/.pypirc.in
+	$(MAKE) -e "template=$(<)" "target=$(@)" expand-template
+
 
 ## Utility Targets:
 #
@@ -437,7 +581,7 @@ $(VCS_FETCH_TARGETS): ./.git/logs/HEAD
 .PHONY: expand-template
 ## Create a file from a template replacing environment variables
 expand-template:
-	$(MAKE) -e "$(HOME)/.local/var/log/project-structure-host-install.log"
+	$(MAKE) -e "$(HOME)/.local/var/log/python-project-structure-host-install.log"
 	set +x
 	if [ -e "$(target)" ]
 	then
@@ -477,8 +621,8 @@ endif
 # the fixed/pinned versions as targets and the `./setup.cfg` file where dependencies are
 # defined as a prerequisite:
 #
-#    ./build/foo.txt: ./foo.txt.in
-#    	envsubst <"$(<)" >"$(@)"
+#    ./requirements.txt: setup.cfg
+#        ./.tox/py310/bin/pip-compile --output-file "$(@)" "$(<)"
 #
 # To that end, developers should use real target files whenever possible when adding
 # recipes to this file.
@@ -491,7 +635,7 @@ endif
 #
 #     ./var/log/foo.log:
 #         mkdir -pv "$(dir $(@))"
-#         echo "Do some work here" | tee -a "$(@)"
+#         ./.tox/build/bin/python "./bin/foo.py" | tee -a "$(@)"
 #
 # This is also useful when none of the modification times of produced artifacts can be
 # counted on to correctly reflect when any subsequent targets need to be updated when
@@ -499,7 +643,7 @@ endif
 # recipe can create arbitrary output:
 #
 #     ./var/log/foo.log:
-#         echo "Do some work here"
+#         ./.tox/build/bin/python "./bin/foo.py"
 #         mkdir -pv "$(dir $(@))"
 #         date | tee -a "$(@)"
 #
